feat: ship sandbox executors in the package
Move both sandbox executors out of examples/ into
pydantic_ai_skills/sandboxes/, mirroring the registries/ layout: the
package __init__ imports both eagerly while each provider SDK is
imported lazily inside a helper that raises an ImportError naming its
extra, exactly as registries/s3.py does. Importing pydantic_ai_skills
never requires an extra to be installed.

OpenSandboxScriptExecutor and LocalSandboxScriptExecutor are now
exported from the top-level package, with a shared _staging module and
an API reference page. The two examples shrink to the same shape as
basic_usage_capability.py — only the script_executor= argument differs.

Because these are library code again, the executor tests come back and
grow: staging boundary, symlink escapes, Pyodide wrapper semantics,
sandbox reuse, and missing-extra errors. 495 tests pass, up from 478.

Also fixes two Codex findings:

- Registry skills could not be sandboxed at all. GitSkillsRegistry and
  S3SkillsRegistry called discover_skills without script_executor and
  exposed no way to set one, so scripts from a cloned repo or synced
  bucket always ran on the host — the exact case docs/sandbox.md holds
  up as the reason to sandbox. Both registries now take script_executor
  and thread it through.

- skill_root_for walked up by script.name depth from a resolved uri.
  Discovery stores an unresolved name but a resolved uri, so an in-tree
  symlink that changes depth (skill/scripts/run.py -> skill/run.py)
  landed on the skills root and staged every sibling skill into the
  sandbox. Reproduced with two skills, confirming skill-b/secret.txt
  became reachable. The root is now anchored on the nearest SKILL.md
  ancestor, with the name-depth walk kept only as a fallback.

Verified end to end against real LocalSandbox: byte-identical output to
local execution for a script using a sibling import, an emoji argument,
a boolean flag, a skill-root resource, stderr and sys.exit(not False).
