+++ docs/api/executors.md
@@ -24,3 +24,23 @@
         heading_level: 2
         members:
             - run
+
+---
+
+::: pydantic_ai_skills.OpenSandboxScriptExecutor
+    options:
+        show_source: true
+        heading_level: 2
+        members:
+            - run
+            - aclose
+
+---
+
+::: pydantic_ai_skills.LocalSandboxScriptExecutor
+    options:
+        show_source: true
+        heading_level: 2
+        members:
+            - run
+            - close
