+++ examples/sandbox_opensandbox.py
@@ -1,302 +1,32 @@
 """Example running skill scripts inside an OpenSandbox container.
 
-This example shows how to create an agent whose skill scripts execute in an
-[OpenSandbox](https://github.com/opensandbox-group/OpenSandbox) container
-rather than a local subprocess. Only the `script_executor=` argument differs
-from `basic_usage_capability.py`.
+This example shows how to create an agent whose skill scripts execute in
+an [OpenSandbox](https://github.com/opensandbox-group/OpenSandbox)
+container rather than a local subprocess. Only the
+`script_executor=` argument differs from `basic_usage_capability.py`.
 
-Requires `pip install -e ".[examples,opensandbox]"` and a reachable OpenSandbox
-server:
+Requires `pip install -e ".[examples,opensandbox]"`.
+
+Also requires a reachable OpenSandbox server:
 
     osb config set connection.domain localhost:8080
     osb config set connection.protocol http
     osb config set connection.api_key <your-api-key>
 
-The executor stages the whole skill folder — `SKILL.md`, `resources/`,
-`scripts/` — and runs the script with its own directory as the working
-directory, so relative paths resolve exactly as they do locally. Symlinks
-resolving outside the skill folder are skipped, so staging cannot pull host
-files into the sandbox. Output is formatted like local execution, so switching
-backends does not change what the model sees.
-
 Note:
-    The bundled `arxiv-search` script imports the `arxiv` package, so it only
-    runs if the container image provides it. The resource-only skills
-    (`pydanticai-docs`, `web-research`) work normally, since resources are read
-    on the host.
+    `arxiv-search` is the only bundled skill with a script, and it needs the
+    `arxiv` package, so it runs only if the container image provides it. The resource-only skills (`pydanticai-docs`,
+    `web-research`) work normally, since resources are read on the host.
 """
 
-from __future__ import annotations
-
-import shlex
-import time
-import warnings
-from collections.abc import Iterator
-from datetime import timedelta
-from pathlib import Path, PurePosixPath
-from typing import TYPE_CHECKING, Any
+from pathlib import Path
 
 import logfire
 import uvicorn
 from dotenv import load_dotenv
 from pydantic_ai import Agent
 
-from pydantic_ai_skills import (
-    LocalSkillScriptExecutor,
-    SkillsCapability,
-    SkillScript,
-    SkillScriptExecutor,
-    SkillsDirectory,
-)
-
-if TYPE_CHECKING:
-    from opensandbox import Sandbox
-
-
-def skill_root_for(script: SkillScript) -> Path:
-    """Resolve the skill folder root that a script belongs to.
-
-    ``script.name`` is relative to the skill folder (for example
-    ``scripts/run.py``), so walking that many levels up from the script file
-    yields the skill root rather than just the script's parent directory.
-
-    Args:
-        script: A file-based script with a ``uri``.
-
-    Returns:
-        Resolved path to the skill folder.
-    """
-    script_path = Path(str(script.uri)).resolve()
-    root = script_path.parent
-    for _ in range(len(PurePosixPath(script.name).parts) - 1):
-        root = root.parent
-    return root
-
-
-def iter_stageable_files(skill_root: Path) -> Iterator[tuple[str, Path]]:
-    """Yield ``(relative_posix_path, resolved_file)`` for files safe to stage.
-
-    Symlinks that resolve outside ``skill_root`` are skipped with a warning.
-    Discovery already rejects those, but staging re-walks the folder, and
-    following such a link would copy an arbitrary host file into the sandbox
-    where the script could read it back out.
-
-    Args:
-        skill_root: Resolved path to the skill folder.
-
-    Yields:
-        Tuples of the path relative to ``skill_root`` and the resolved file.
-    """
-    for path in sorted(skill_root.rglob('*')):
-        resolved = path.resolve()
-        if not resolved.is_relative_to(skill_root):
-            warnings.warn(
-                f"Skipping '{path}': resolves outside the skill folder (symlink escape detected).",
-                UserWarning,
-                stacklevel=2,
-            )
-            continue
-        if resolved.is_file():
-            yield path.relative_to(skill_root).as_posix(), resolved
-
-
-def _stage_snapshot(skill_root: Path) -> tuple[list[tuple[str, Path]], tuple[tuple[str, int, int], ...]]:
-    """Walk the skill folder once, returning its files and a change fingerprint.
-
-    The fingerprint covers every staged path with its size and modification
-    time, so a reused sandbox can tell whether the source skill was edited since
-    it was staged — ``auto_reload`` and ``reload()`` both surface edits under an
-    unchanged skill root.
-
-    Args:
-        skill_root: Resolved path to the skill folder.
-
-    Returns:
-        The staged ``(relative, resolved)`` pairs and their fingerprint.
-    """
-    entries: list[tuple[str, Path]] = []
-    fingerprint: list[tuple[str, int, int]] = []
-    for relative, resolved in iter_stageable_files(skill_root):
-        entries.append((relative, resolved))
-        stat = resolved.stat()
-        fingerprint.append((relative, stat.st_size, stat.st_mtime_ns))
-    return entries, tuple(fingerprint)
-
-
-# Suffix -> interpreter, resolved inside the sandbox rather than on the host.
-_SANDBOX_INTERPRETERS: dict[str, list[str]] = {
-    '.py': ['python3'],
-    '.sh': ['sh'],
-    '.bash': ['bash'],
-    '.zsh': ['zsh'],
-    '.fish': ['fish'],
-}
-
-
-def _require_opensandbox() -> Any:
-    """Import the opensandbox SDK, or explain which extra installs it."""
-    try:
-        from opensandbox import Sandbox as _Sandbox
-    except ImportError as exc:  # pragma: no cover - depends on the environment
-        raise ImportError(
-            'OpenSandboxScriptExecutor requires the "opensandbox" package. '
-            'Install it with: pip install "pydantic-ai-skills[opensandbox]"'
-        ) from exc
-
-    return _Sandbox
-
-
-class OpenSandboxScriptExecutor(SkillScriptExecutor):
-    """Execute file-based skill scripts inside an OpenSandbox container.
-
-    Attributes:
-        timeout: Per-script execution timeout in seconds.
-    """
-
-    def __init__(
-        self,
-        image: str = 'opensandbox/code-interpreter:v1.1.0',
-        *,
-        timeout: int = 30,
-        workdir: str = '/tmp/skills',
-        env_vars: dict[str, str] | None = None,
-        reuse_sandbox: bool = False,
-        sandbox_timeout: timedelta = timedelta(minutes=10),
-    ) -> None:
-        """Initialize the OpenSandbox executor.
-
-        Args:
-            image: Container image used for each sandbox.
-            timeout: Per-script execution timeout in seconds.
-            workdir: Directory inside the sandbox that the skill folder is staged into.
-            env_vars: Environment variables exported to the script process.
-            reuse_sandbox: Keep a single sandbox alive across runs instead of
-                creating and killing one per run. Faster, but runs share state.
-            sandbox_timeout: Lifetime of the sandbox itself, passed to ``Sandbox.create``.
-        """
-        self.timeout = timeout
-        self._image = image
-        self._workdir = workdir.rstrip('/') or '/tmp/skills'
-        self._env_vars = dict(env_vars or {})
-        self._reuse_sandbox = reuse_sandbox
-        self._sandbox_timeout = sandbox_timeout
-        self._sandbox: Sandbox | None = None
-        self._sandbox_deadline: float = 0.0
-        # Reused for its host-independent argument marshalling and output formatting.
-        self._formatter = LocalSkillScriptExecutor()
-
-    async def _get_sandbox(self) -> Sandbox:
-        """Return the sandbox to run in, creating one when needed.
-
-        A reused sandbox is replaced once its server-side lifetime is close to
-        expiring. ``Sandbox.create`` fixes that lifetime, so holding the handle
-        past it would send every later run to an expired sandbox. The deadline
-        leaves one script timeout of headroom so a run started now can finish.
-        """
-        if self._reuse_sandbox and self._sandbox is not None:
-            if time.monotonic() < self._sandbox_deadline:
-                return self._sandbox
-            await self.aclose()
-
-        sandbox_cls = _require_opensandbox()
-        sandbox: Sandbox = await sandbox_cls.create(
-            self._image,
-            env=self._env_vars or None,
-            timeout=self._sandbox_timeout,
-        )
-        if self._reuse_sandbox:
-            self._sandbox = sandbox
-            self._sandbox_deadline = time.monotonic() + self._sandbox_timeout.total_seconds() - self.timeout
-        return sandbox
-
-    async def _stage_skill_folder(self, sandbox: Sandbox, skill_root: Path) -> None:
-        """Upload every stageable file in the skill folder into the sandbox workdir."""
-        from opensandbox.models import WriteEntry
-
-        entries, _ = _stage_snapshot(skill_root)
-        write_entries = [
-            WriteEntry(
-                path=f'{self._workdir}/{relative}',
-                data=resolved.read_bytes(),
-                mode=0o755 if resolved.stat().st_mode & 0o111 else 0o644,
-            )
-            for relative, resolved in entries
-        ]
-
-        if write_entries:
-            await sandbox.files.write_files(write_entries)
-
-    def _build_command(self, remote_path: str, suffix: str, args: dict[str, Any] | None) -> str:
-        """Build the shell command line executed inside the sandbox."""
-        interpreter = _SANDBOX_INTERPRETERS.get(suffix)
-        cmd = [*interpreter, remote_path] if interpreter else [remote_path]
-
-        if args:
-            # Reuse the built-in bool/list/None marshalling rules.
-            self._formatter._build_args(cmd, args)
-
-        return shlex.join(cmd)
-
-    async def run(
-        self,
-        script: SkillScript,
-        args: dict[str, Any] | None = None,
-        ctx: Any | None = None,
-    ) -> Any:
-        """Run a skill script inside an OpenSandbox container.
-
-        Args:
-            script: The script to run; ``script.uri`` must point at a local file.
-            args: Named arguments, marshalled with the same rules as
-                [`LocalSkillScriptExecutor`][pydantic_ai_skills.LocalSkillScriptExecutor].
-            ctx: Unused; accepted for protocol compatibility.
-
-        Returns:
-            Combined stdout and stderr, formatted like local execution.
-
-        Raises:
-            ValueError: If the script has no URI configured.
-        """
-        if script.uri is None:
-            raise ValueError(f"Script '{script.name}' has no URI for sandbox execution")
-
-        script_path = Path(script.uri).resolve()
-        skill_root = skill_root_for(script)
-        remote_path = f'{self._workdir}/{script_path.relative_to(skill_root).as_posix()}'
-        # cwd is the script's own directory, matching LocalSkillScriptExecutor.
-        working_directory = str(PurePosixPath(remote_path).parent)
-        command = self._build_command(remote_path, script_path.suffix.lower(), args)
-
-        # _get_sandbox raises the ImportError naming the extra, so import the SDK
-        # models only once a sandbox exists.
-        sandbox = await self._get_sandbox()
-        from opensandbox.models.execd import RunCommandOpts
-
-        try:
-            await self._stage_skill_folder(sandbox, skill_root)
-            execution = await sandbox.commands.run(
-                command,
-                opts=RunCommandOpts(
-                    working_directory=working_directory,
-                    timeout=timedelta(seconds=self.timeout),
-                    envs=self._env_vars or None,
-                ),
-            )
-        finally:
-            if not self._reuse_sandbox:
-                await sandbox.kill()
-
-        stdout = ''.join(message.text for message in execution.logs.stdout)
-        stderr = ''.join(message.text for message in execution.logs.stderr)
-        return self._formatter._format_output([stdout.encode()], [stderr.encode()], execution.exit_code or 0)
-
-    async def aclose(self) -> None:
-        """Kill the reused sandbox, if one is alive."""
-        if self._sandbox is not None:
-            await self._sandbox.kill()
-            self._sandbox = None
-            self._sandbox_deadline = 0.0
-
+from pydantic_ai_skills import OpenSandboxScriptExecutor, SkillsCapability, SkillsDirectory
 
 load_dotenv()
