+++ pydantic_ai_skills/registries/s3.py
@@ -14,6 +14,7 @@ from pathlib import Path
 from typing import Any
 
 from pydantic_ai_skills.directory import discover_skills
+from pydantic_ai_skills.executors import SkillScriptExecutor
 from pydantic_ai_skills.registries._base import SkillRegistry
 from pydantic_ai_skills.registries._copy import copy_skill_directory
 from pydantic_ai_skills.types import Skill
@@ -63,6 +64,11 @@ class S3SkillsRegistry(SkillRegistry):
             ``False`` to skip syncing on construction and on ``search`` / ``get``;
             the registry then reads only what already exists in ``target_dir``.
             Note that an explicit ``update()`` call always contacts S3.
+        script_executor: Executor used for file-based scripts in the discovered
+            skills. Registry skills are the least-trusted source there is, so
+            pass a sandboxing executor here to keep them off the host. When
+            None, scripts run as local subprocesses via
+            :class:`~pydantic_ai_skills.LocalSkillScriptExecutor`.
 
     Examples:
         Amazon S3 with the ambient credential chain:
@@ -103,6 +109,7 @@ class S3SkillsRegistry(SkillRegistry):
         boto3_client: Any | None = None,
         validate: bool = True,
         auto_install: bool = True,
+        script_executor: SkillScriptExecutor | None = None,
     ) -> None:
         if boto3_client is None:
             try:
@@ -120,6 +127,7 @@ class S3SkillsRegistry(SkillRegistry):
         self._bucket = bucket
         self._prefix = prefix.strip('/')
         self._validate = validate
+        self._script_executor = script_executor
         self._auto_install = auto_install
         self._tmp_dir: Any | None = None
         # Cache of the most recent object listing (Key -> LastModified), populated by _sync.
@@ -210,7 +218,12 @@ class S3SkillsRegistry(SkillRegistry):
         skills_root = self._skills_root()
         if not skills_root.exists():
             return []
-        return discover_skills(path=skills_root, validate=self._validate, max_depth=2)
+        return discover_skills(
+            path=skills_root,
+            validate=self._validate,
+            max_depth=2,
+            script_executor=self._script_executor,
+        )
 
     def _skill_version(self, skill: Skill) -> str | None:
         """Return the latest LastModified across the skill's objects, ISO-formatted.
